pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-optimizer-arab-backend"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109,<0.112",
    "uvicorn[standard]>=0.27,<0.30",
    "SQLAlchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "httpx[http2]>=0.26,<0.28",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.1,<3.0",
    "python-dotenv>=1.0,<2.0",
    "psycopg[binary]>=3.1,<4.0",
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<0.25",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "supabase>=2.0,<3.0",
    "boto3>=1.34,<2.0",
    "Pillow>=10.0,<12.0",
    "python-multipart>=0.0.6,<0.1.0",
    # Cross-service validation cache
    "redis>=5.0,<6.0",
    # Rate limiting
    "slowapi>=0.1.9,<0.2.0",
    "openai>=1.0,<2",
    "google-genai",
    "anthropic>=0.40,<1.0",
    "numpy>=1.26,<2.0",
    "pymupdf>=1.24,<1.25",
    "python-docx>=1.1,<2.0",
    "easyocr>=1.7,<2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2,<0.5",
]

[tool.setuptools.packages.find]
include = ["libs*", "services*", "alembic*"]

[tool.ruff]
extend-exclude = [
    "alembic/versions",
    "services/*/alembic/versions",
]

[tool.ruff.lint]
ignore = ["E402"]
